Move pkg/kubectl/apply.go to staging
