+++ pkg/kubectl/cmd/util/editor/BUILD
@@ -15,7 +15,6 @@ go_library(
         "//build/visible_to:pkg_kubectl_cmd_util_editor_CONSUMERS",
     ],
     deps = [
-        "//pkg/kubectl:go_default_library",
         "//pkg/kubectl/cmd/util:go_default_library",
         "//pkg/kubectl/cmd/util/editor/crlf:go_default_library",
         "//staging/src/k8s.io/api/core/v1:go_default_library",
@@ -33,6 +32,7 @@ go_library(
         "//staging/src/k8s.io/cli-runtime/pkg/printers:go_default_library",
         "//staging/src/k8s.io/cli-runtime/pkg/resource:go_default_library",
         "//staging/src/k8s.io/kubectl/pkg/scheme:go_default_library",
+        "//staging/src/k8s.io/kubectl/pkg/util:go_default_library",
         "//staging/src/k8s.io/kubectl/pkg/util/term:go_default_library",
         "//vendor/github.com/evanphx/json-patch:go_default_library",
         "//vendor/github.com/spf13/cobra:go_default_library",
